+++ example/lib/barcode_scanner_window.dart
@@ -21,77 +21,6 @@ class _BarcodeScannerWithScanWindowState
   // TODO: Fix BoxFit.fill & BoxFit.fitHeight
   final boxFit = BoxFit.contain;
 
-  Widget _buildBarcodeOverlay() {
-    return ValueListenableBuilder(
-      valueListenable: controller,
-      builder: (context, value, child) {
-        // Not ready.
-        if (!value.isInitialized || !value.isRunning || value.error != null) {
-          return const SizedBox();
-        }
-
-        return StreamBuilder<BarcodeCapture>(
-          stream: controller.barcodes,
-          builder: (context, snapshot) {
-            final BarcodeCapture? barcodeCapture = snapshot.data;
-
-            // No barcode.
-            if (barcodeCapture == null || barcodeCapture.barcodes.isEmpty) {
-              return const SizedBox();
-            }
-
-            final overlays = <Widget>[];
-
-            for (final scannedBarcode in barcodeCapture.barcodes) {
-              // No barcode corners, or size, or no camera preview size.
-              if (value.size.isEmpty ||
-                  scannedBarcode.size.isEmpty ||
-                  scannedBarcode.corners.isEmpty) {
-                continue;
-              }
-
-              overlays.add(
-                CustomPaint(
-                  painter: BarcodeOverlay(
-                    barcodeCorners: scannedBarcode.corners,
-                    barcodeSize: scannedBarcode.size,
-                    boxFit: boxFit,
-                    cameraPreviewSize: barcodeCapture.size,
-                  ),
-                ),
-              );
-            }
-
-            return Stack(
-              fit: StackFit.expand,
-              children: overlays,
-            );
-          },
-        );
-      },
-    );
-  }
-
-  Widget _buildScanWindow(Rect scanWindowRect) {
-    return ValueListenableBuilder(
-      valueListenable: controller,
-      builder: (context, value, child) {
-        // Not ready.
-        if (!value.isInitialized ||
-            !value.isRunning ||
-            value.error != null ||
-            value.size.isEmpty) {
-          return const SizedBox();
-        }
-
-        return CustomPaint(
-          size: value.size,
-          painter: ScannerOverlay(scanWindowRect),
-        );
-      },
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     final scanWindow = Rect.fromCenter(
@@ -114,8 +43,11 @@ class _BarcodeScannerWithScanWindowState
               return ScannerErrorWidget(error: error);
             },
           ),
-          _buildBarcodeOverlay(),
-          _buildScanWindow(scanWindow),
+          BarcodeOverlay(controller: controller, boxFit: boxFit),
+          ScannerOverlay(
+            scanWindow: scanWindow,
+            controller: controller,
+          ),
           Align(
             alignment: Alignment.bottomCenter,
             child: Container(
@@ -135,107 +67,5 @@ class _BarcodeScannerWithScanWindowState
   Future<void> dispose() async {
     super.dispose();
     await controller.dispose();
-  }
-}
-
-class ScannerOverlay extends CustomPainter {
-  ScannerOverlay(this.scanWindow);
-
-  final Rect scanWindow;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    // Define the main overlay path covering the entire screen
-    final backgroundPath = Path()
-      ..addRect(Rect.fromLTWH(0, 0, size.width, size.height));
-
-    // Define the cutout path in the center
-    final cutoutPath = Path()..addRect(scanWindow);
-
-    // Combine the two paths: overlay minus the cutout area
-    final overlayWithCutoutPath =
-        Path.combine(PathOperation.difference, backgroundPath, cutoutPath);
-
-    // Paint the overlay with the cutout
-    final paint = Paint()
-      ..color = Colors.black.withOpacity(0.5); // Semi-transparent black
-    canvas.drawPath(overlayWithCutoutPath, paint);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return false;
-  }
-}
-
-class BarcodeOverlay extends CustomPainter {
-  BarcodeOverlay({
-    required this.barcodeCorners,
-    required this.barcodeSize,
-    required this.boxFit,
-    required this.cameraPreviewSize,
-  });
-
-  final List<Offset> barcodeCorners;
-  final Size barcodeSize;
-  final BoxFit boxFit;
-  final Size cameraPreviewSize;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    if (barcodeCorners.isEmpty ||
-        barcodeSize.isEmpty ||
-        cameraPreviewSize.isEmpty) {
-      return;
-    }
-
-    final adjustedSize = applyBoxFit(boxFit, cameraPreviewSize, size);
-
-    double verticalPadding = size.height - adjustedSize.destination.height;
-    double horizontalPadding = size.width - adjustedSize.destination.width;
-    if (verticalPadding > 0) {
-      verticalPadding = verticalPadding / 2;
-    } else {
-      verticalPadding = 0;
-    }
-
-    if (horizontalPadding > 0) {
-      horizontalPadding = horizontalPadding / 2;
-    } else {
-      horizontalPadding = 0;
-    }
-
-    final double ratioWidth;
-    final double ratioHeight;
-
-    // if (!kIsWeb && defaultTargetPlatform == TargetPlatform.iOS) {
-    //   ratioWidth = barcodeSize.width / adjustedSize.destination.width;
-    //   ratioHeight = barcodeSize.height / adjustedSize.destination.height;
-    // } else
-    // {
-    ratioWidth = cameraPreviewSize.width / adjustedSize.destination.width;
-    ratioHeight = cameraPreviewSize.height / adjustedSize.destination.height;
-    // }
-
-    final List<Offset> adjustedOffset = [
-      for (final offset in barcodeCorners)
-        Offset(
-          offset.dx / ratioWidth + horizontalPadding,
-          offset.dy / ratioHeight + verticalPadding,
-        ),
-    ];
-
-    final cutoutPath = Path()..addPolygon(adjustedOffset, true);
-
-    final backgroundPaint = Paint()
-      ..color = Colors.red.withOpacity(0.3)
-      ..style = PaintingStyle.fill;
-
-    canvas.drawPath(cutoutPath, backgroundPaint);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return true;
   }
 }